+++ test/test_processes.jl
@@ -57,7 +57,7 @@ M = rand(2)
 Sigma = [0.5 0.5; 0.5 0.5]
 R = [0.9 0.1; 0 .8];
 n_states= [5,10]
-n_states::Array{Int64,1}
+n_states::Array{Int,1}
 n = 2
 n_integration=[7,4]
 # 4/ Create a VAR1 type (in the module). Add new functions that operate on this type. Add them to the test file.
@@ -77,7 +77,7 @@ M = rand(1)
 Sigma = [0.5]'
 R = [0.9]'
 n_states= [5]
-n_states::Array{Int64,1}
+n_states::Array{Int,1}
 n = 2
 n_integration=[7]
 # 4/ Create a VAR1 type (in the module). Add new functions that operate on this type. Add them to the test file.
